+++ pkg/contracts/testdappv2/TestDAppV2.sol
@@ -13,10 +13,12 @@ contract TestDAppV2 {
     }
 
     /// @notice Struct containing revert context passed to onRevert.
+    /// @param sender Address of account that initiated smart contract call.
     /// @param asset Address of asset, empty if it's gas token.
     /// @param amount Amount specified with the transaction.
     /// @param revertMessage Arbitrary data sent back in onRevert.
     struct RevertContext {
+        address sender;
         address asset;
         uint64 amount;
         bytes revertMessage;
@@ -100,6 +102,7 @@ contract TestDAppV2 {
     function onRevert(RevertContext calldata revertContext) external {
         setCalledWithMessage(string(revertContext.revertMessage));
         setAmountWithMessage(string(revertContext.revertMessage), 0);
+        senderWithMessage[revertContext.revertMessage] = revertContext.sender;
     }
 
     function setExpectedOnCallSender(address _expectedOnCallSender) external {